Initial Commit
Most of this code was copied from my other repository named
`shell_base`.

I fixed a *lot* of typos and totally reworked how the code is
organized. I did this because the way it's done in `shell_base`
is just way too complicated. Plus I didn't document it well,
so I don't even understand how it's supposed to work.

Note that there is a known issue in this committed code.
Specifically that `command -v <someFunctionName>` doesn't
work when called from `main.sh`.